% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossmap.R
\name{crossmap_bed}
\alias{crossmap_bed}
\title{Crossmap BED file to different assembly version}
\usage{
crossmap_bed(
  direction = "hg38Tohg19",
  crossmap_cmd_path = "/Users/sigven/miniconda3/envs/py36/bin/CrossMap.py",
  chain_file_directory = "/Users/sigven/research/DB/chainFiles",
  wdir = NULL,
  debug = FALSE,
  fsep = "_",
  source_bed = NULL,
  target_bed = NULL
)
}
\arguments{
\item{direction}{hg38Tohg19 or hg19Tohg38}

\item{crossmap_cmd_path}{Full path to CrossMap.py file}

\item{chain_file_directory}{Directory containing chain files}

\item{wdir}{working directory}

\item{debug}{logical indicating if debug mode is on or off}

\item{fsep}{file name separator (default '_')}

\item{source_bed}{BED file to be lifted}

\item{target_bed}{BED file with lifted coordinates}
}
\description{
Function that "lifts" the BED records to a different
assembly version of the human genome, using CrossMap.py
}
