% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_bed_records}
\alias{order_bed_records}
\title{Sort BED segments}
\usage{
order_bed_records(
  bed_records,
  chrom_col = "chrom",
  start_col = "start",
  end_col = "end"
)
}
\arguments{
\item{bed_records}{data frame with BED segment records}

\item{chrom_col}{name of chromosome column}

\item{start_col}{name of start column}

\item{end_col}{name of end column}
}
\description{
Sort BED segments
}
