% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_vcf_records}
\alias{order_vcf_records}
\title{Order VCF data frame records}
\usage{
order_vcf_records(vcf_records = NULL, chrom_var = "CHROM", pos_var = "POS")
}
\arguments{
\item{vcf_records}{data frame with VCF records}

\item{chrom_var}{variable name in vcf_df for chromosome}

\item{pos_var}{variable name in vcf_df for chromosomal position}
}
\value{
vcf_df data frame with ordered VCF records
}
\description{
Function that orders VCF records according to order
of chromosomes and chromosomal position
}
