% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_bed_records}
\alias{write_bed_records}
\title{Write BED data frame records}
\usage{
write_bed_records(
  bed_records = NULL,
  otput_dir = NA,
  bed_fname_prefix = NA,
  genome_build = "grch37",
  fseparator = "_",
  chrom_col = "chrom",
  start_col = "start",
  end_col = "end",
  keep_uncompressed = F
)
}
\arguments{
\item{bed_records}{data frame with column names and data similar to
mandatory BED columns (CHROM, START, END etc)}

\item{bed_fname_prefix}{VCF file name prefix}

\item{genome_build}{human genome build (grch37/grch38)}

\item{fseparator}{separator symbol to use in filenames}

\item{chrom_col}{name of chromosome column}

\item{start_col}{name of start column}

\item{end_col}{name of end column}

\item{keep_uncompressed}{logical indicating if uncompressed BED file should
be kept}

\item{output_dir}{output directory for VCF file}
}
\description{
Sort and write BED records to file
}
