% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_vcf_records}
\alias{write_vcf_records}
\title{Write VCF data frame records}
\usage{
write_vcf_records(
  vcf_records = NULL,
  output_dir = NULL,
  header_lines = NULL,
  vcf_fname_prefix = NULL,
  fseparator = "_",
  genome_build = "grch38",
  sample_names = NULL,
  validate = F
)
}
\arguments{
\item{vcf_records}{data frame with column names and data similar to mandatory VCF columns (CHROM, POS, ID etc)}

\item{output_dir}{output directory for VCF file}

\item{header_lines}{character vector with VCF header lines}

\item{vcf_fname_prefix}{VCF file name prefix}

\item{fseparator}{separator symbol to use in filenames}

\item{genome_build}{human genome build (grch37/grch38)}

\item{sample_names}{sample names for genotype columns}

\item{validate}{logical indicating if vcf-validator should be run on final VCF file}
}
\description{
Validate, sort and write VCF records to file
}
